### 2024-10-14 #7348 @editorialbot
[PRE REVIEW]: AliNe: A Flexible and Efficient Nextflow Pipeline for Read Alignmen
**Submitting author:** @Juke34 (<a href="http://orcid.org/0000-0002-6629-0173">Jacques Dainat</a>)
**Repository:** https://github.com/Juke34/AliNe
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
