### 2024-10-20 #7348 @Kevin-Mattheus-Moerman
- [ ] Since `cloc` does not count lines of nextflow code, could you help provide a summary of the number of lines of code contained in this submission (that are fully your own work). Thanks!
- [ ] Please format the paper as per our guidelines, i.e. a `Statement of Need` section is needed. You can call `@editorialbot generate pdf` here to update the paper.
