### 2024-10-21 f524554
fix doi and tophat2 mispelling
