+++ paper/paper.bib
@@ -74,10 +74,10 @@
    url = {https://pubmed.ncbi.nlm.nih.gov/23558742/},
    year = {2013},
 }
-@article{topha2,
+@article{tophat2,
    abstract = {TopHat is a popular spliced aligner for RNA-sequence (RNA-seq) experiments. In this paper, we describe TopHat2, which incorporates many significant enhancements to TopHat. TopHat2 can align reads of various lengths produced by the latest sequencing technologies, while allowing for variable-length indels with respect to the reference genome. In addition to de novo spliced alignment, TopHat2 can align reads across fusion breaks, which can occur after genomic translocations. TopHat2 combines the ability to identify novel splice sites with direct mapping to known transcripts, producing sensitive and accurate alignments, even for highly repetitive genomes or in the presence of pseudogenes. TopHat2 is available at http://ccb.jhu.edu/software/tophat. © 2013 Kim et al.; licensee BioMed Central Ltd.},
    author = {Daehwan Kim and Geo Pertea and Cole Trapnell and Harold Pimentel and Ryan Kelley and Steven L. Salzberg},
-   doi = {10.1186/GB-2013-14-4-R36/FIGURES/6},
+   doi = {10.1186/GB-2013-14-4-R36},
    issn = {1474760X},
    issue = {4},
    journal = {Genome Biology},
