+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 Alignment of sequencing reads against a reference genome is a critical step in many bioinformatics workflows, especially in genomics, transcriptomics, and population genetics. Different aligners offer varying performance characteristics in terms of speed, memory efficiency, and accuracy, which makes the selection of an appropriate tool context-dependent. To streamline this process, we present AliNe (Alignment in Nextflow), a flexible and efficient read alignment pipeline built on the Nextflow framework [@nextflow], designed to accommodate a range of commonly used aligners in high-throughput sequencing projects.
 
-AliNe supports short reads (both paired-end and single-end) as well as long reads generated by PacBio and Oxford Nanopore Technologies (ONT). It currently supports 15 widely used alignment tools, including BBMap [@bbmap], Bowtie2 [@bowtie2], BWA [@bwaaln], BWA-MEM [@bwamem], BWA-SW [@bwasw], GraphMap2 [@graphmap2], HISAT2 [@hisat2], Minimap2 [@minimap2], ngmlr [@ngmlr], novoAlign [@novoalign], nucmer [@nucmer], STAR (single or two-pass mode) [@star], subread [@subread], sublong [@subread] and Tophat2 [tophat2]. These aligners are integrated into a single, easy-to-use workflow, giving users the flexibility to choose the best tool for their specific data and objectives.
+AliNe supports short reads (both paired-end and single-end) as well as long reads generated by PacBio and Oxford Nanopore Technologies (ONT). It currently supports 15 widely used alignment tools, including BBMap [@bbmap], Bowtie2 [@bowtie2], BWA [@bwaaln], BWA-MEM [@bwamem], BWA-SW [@bwasw], GraphMap2 [@graphmap2], HISAT2 [@hisat2], Minimap2 [@minimap2], ngmlr [@ngmlr], novoAlign [@novoalign], nucmer [@nucmer], STAR (single or two-pass mode) [@star], subread [@subread], sublong [@subread] and Tophat2 [@tophat2]. These aligners are integrated into a single, easy-to-use workflow, giving users the flexibility to choose the best tool for their specific data and objectives.
 
 AliNe is designed to minimize user inputs and avoid common mistakes like the definition of scoring system used by the sequencing machine (Phred+33, Phred+64 and Solexa), and for RNAseq data the strandness (stranded or unstranded) or the orientation in case of paired-end reads (forward reverse i.e., fr; reverse forward i.e., rf; forward forward i.e., ff). This information may impact the results. If not provided by the user, AliNe will deduce this information and adapt the option of the different downstream aligners accordingly.
 
