+++ paper/paper.bib
@@ -91,14 +91,15 @@
    url = {https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-4-r36},
    year = {2013},
 }
-@article{bwamem,
-   abstract = {Summary: BWA-MEM is a new alignment algorithm for aligning sequence reads or long query sequences against a large reference genome such as human. It automatically chooses between local and end-to-end alignments, supports paired-end reads and performs chimeric alignment. The algorithm is robust to sequencing errors and applicable to a wide range of sequence lengths from 70bp to a few megabases. For mapping 100bp sequences, BWA-MEM shows better performance than several state-of-art read aligners to date. Availability and implementation: BWA-MEM is implemented as a component of BWA, which is available at http://github.com/lh3/bwa. Contact: hengli@broadinstitute.org},
-   author = {Heng Li},
-   keywords = {()},
-   month = {3},
-   title = {Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM},
-   url = {https://arxiv.org/abs/1303.3997v2},
-   year = {2013},
+@misc{bwamem,
+      title={Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM}, 
+      author={Heng Li},
+      abstract = {Summary: BWA-MEM is a new alignment algorithm for aligning sequence reads or long query sequences against a large reference genome such as human. It automatically chooses between local and end-to-end alignments, supports paired-end reads and performs chimeric alignment. The algorithm is robust to sequencing errors and applicable to a wide range of sequence lengths from 70bp to a few megabases. For mapping 100bp sequences, BWA-MEM shows better performance than several state-of-art read aligners to date. Availability and implementation: BWA-MEM is implemented as a component of BWA, which is available at http://github.com/lh3/bwa. Contact: hengli@broadinstitute.org},
+      year={2013},
+      eprint={1303.3997},
+      archivePrefix={arXiv},
+      primaryClass={q-bio.GN},
+      url={https://arxiv.org/abs/1303.3997}, 
 }
 @article{star,
    abstract = {Motivation: Accurate alignment of high-throughput RNA-seq data is a challenging and yet unsolved problem because of the non-contiguous transcript structure, relatively short read lengths and constantly increasing throughput of the sequencing technologies. Currently available RNA-seq aligners suffer from high mapping error rates, low mapping speed, read length limitation and mapping biases.Results: To align our large (>80 billon reads) ENCODE Transcriptome RNA-seq dataset, we developed the Spliced Transcripts Alignment to a Reference (STAR) software based on a previously undescribed RNA-seq alignment algorithm that uses sequential maximum mappable seed search in uncompressed suffix arrays followed by seed clustering and stitching procedure. STAR outperforms other aligners by a factor of >50 in mapping speed, aligning to the human genome 550 million 2 × 76 bp paired-end reads per hour on a modest 12-core server, while at the same time improving alignment sensitivity and precision. In addition to unbiased de novo detection of canonical junctions, STAR can discover non-canonical splices and chimeric (fusion) transcripts, and is also capable of mapping full-length RNA sequences. Using Roche 454 sequencing of reverse transcription polymerase chain reaction amplicons, we experimentally validated 1960 novel intergenic splice junctions with an 80-90% success rate, corroborating the high precision of the STAR mapping strategy. © The Author(s) 2012. Published by Oxford University Press.},
@@ -221,9 +222,21 @@
    year = {2019},
 }
 @article{hisat2,
-   author = {Ryan J Musich},
-   title = {A Recent ( 2020 ) Comparative Analysis of Genome Aligners Shows HISAT2 and BWA are Among the Best Tools by},
-   year = {2020},
[…]
