### 2024-10-21 #7348 @Juke34
Dear @Kevin-Mattheus-Moerman,

I have addressed all the issues you pointed out.
About the number of lines of code contained in this submission, it should be 3538 lines (counted using `wc -l`)

Best regards,

Jacques Dainat
