### 2024-11-13 #7348 @jromanowska
> I can review!
>
> I am a cis male, though, so not sure whether I'm still eligible 😅

Thanks, @rcannood! Yes, you are eligible. ;) I was asking Juliette for recommendations for someone non cis-male because all the other people I could find were (assumingly) cis-male. It's always good to have a broad spectrum of reviewers, if possible.
