### 2025-01-06 966665a
fix typos, add links
