### 2025-01-06 Juke34/AliNe#2 @rcannood
JOSS review: Improvements to the README
Hi @Juke34 ! I'm reviewing the AliNe JOSS submission in https://github.com/openjournals/joss-reviews/issues/7545#issuecomment-2529060329.

Here are some proposed improvements for the README of the repository:

# Overall

* [x] Read through the README to fix grammatical and spelling errors
* [x] Fix markdown rendering issues (PR #1 fixes the `** Table1**` not rendering properly and adds links to the tables

# Foreword

* [x] Make the tables more legible: The table currently consists of `X recomended`and `x Not recommended` which is hard to distinguish between. Perhaps something like: ✅ Supported, ⚠️ Not recommended, 🚫 Not applicable.

# Flowchart

* [x] Make the mermaid chart more legible. In dark mode, the cross hatching makes the mermaid chart unnecessarily hard to read.

# Installation & Usage

* [x] Rather than clone the repository, you should let Nextflow fetch the workflow from the github repository. That is, rather than:

```bash
git clone https://github.com/Juke34/AliNe.git
cd AliNe
nextflow run aline.nf ...
```

You should instruct your users to run the workflow directly from the repository:

```bash
nextflow run Juke34/AliNe -r v1.0.0 # ...
```

For that you will need to rename `aline.nf` to `main.nf` or Nextflow won't be able to detect the workflow correctly.
