### 2025-01-06 Juke34/AliNe#3 @rcannood
JOSS review: Improvements to the software paper
I'm reviewing the JOSS submission of AliNe in https://github.com/openjournals/joss-reviews/issues/7545#issuecomment-2529060329. This issue proposes some changes for the software paper.

# Summary

> Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary is currently a bit too long and not high-level enough. For example, it has a list of all of the different supported and detailed information on parameters (scoring system, standedness, orientation). I would move some of this information to a separate section related to "Workflow Overview"

# Statement of need

> Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

This is mostly ok, though I feel the target audience and use case is a bit too vague. Currently:

> ... users often need to evaluate multiple aligners for optimal results or based on specific project requirements. By integrating multiple aligners into a unified pipeline, AliNe simplifies this decision-making process and provides a standard workflow to ensure reproducibility and transparency across analyses.

Can the authors demonstrate examples of where AliNe is used to simplify the decision-making process, or showcase a documentation vignette of how a user would do this using AliNe on a demo dataset?

# State of the field

> Do the authors describe how this software compares to other commonly-used packages?

This section is missing.

# Quality of writing

> Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

In terms of quality of writing, everything looks good! However the "Workflow overview" section does not provide enough information for me to immediately understand what AliNe is, what it does and how it works. I feel like some of the information from the "Summary" section should be moved here, and some of the information in the README of the repository should also be moved here (e.g. the mermaid diagram or the tables).

# References

> Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

The BBMap reference needs to be fixed -- similar to how FastQC is rendered I guess.
