### 2025-01-06 #7545 @rcannood
> @rcannood , @gchure - I hope you had a great end-of-the-year break. Just sending a kind reminder about this review. Please update us on your progress as soon as you start working on it.

Thanks for the reminder! I started work on reviewing the functionality and software paper.
