### 2025-01-06 66808d3
Update explanation for library type parameters
