### 2025-01-06 5e5d1d0
tell user to use Aline directly from repo and version: nextflow run Juke34/AliNe -r v1.0.0
