### 2025-01-06 Juke34/AliNe#2 @Juke34
Thank you for all these valuables feedbacks.
Btw for the last recommendation no need to rename aline.nf into main.nf
