### 2025-01-06 92e79d8
remove uninstall section
