### 2025-01-06 #7545 @gchure
> @rcannood , @gchure - I hope you had a great end-of-the-year break. Just sending a kind reminder about this review. Please update us on your progress as soon as you start working on it.

Thanks for the reminder as well! My plan is to resume working on it today.

Happy New Year to everyone!
