+++ paper/paper.bib
@@ -1,11 +1,137 @@
+@software{tronflow,
+  author       = {Pablo Riesgo-Ferreiro and
+                  tronbfxdev and
+                  Julian Mohr},
+  title        = {TRON-Bioinformatics/tronflow-alignment: v1.8.0},
+  month        = oct,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {v1.8.0},
+  doi          = {10.5281/zenodo.7219617},
+  url          = {https://doi.org/10.5281/zenodo.7219617},
+}
+@software{nfcorernaseq,
+  author       = {Harshil Patel and
+                  Phil Ewels and
+                  Jonathan Manning and
+                  Maxime U Garcia and
+                  Alexander Peltzer and
+                  Rickard Hammarén and
+                  Olga Botvinnik and
+                  Adam Talbot and
+                  Gregor Sturm and
+                  nf-core bot and
+                  Matthias Zepper and
+                  Denis Moreno and
+                  Pranathi Vemuri and
+                  Mahesh Binzer-Panchal and
+                  Ezra Greenberg and
+                  silviamorins and
+                  Lorena Pantano and
+                  Robert Syme and
+                  Gavin Kelly and
+                  Friederike Hanssen and
+                  James A. Fellows Yates and
+                  Jose Espinosa-Carrasco and
+                  rfenouil and
+                  Luke Zappia and
+                  Chris Cheshire and
+                  Edmund Miller and
+                  marchoeppner and
+                  Peng Zhou and
+                  Sarah Guinchard and
+                  Gisela Gabernet},
+  title        = {nf-core/rnaseq: nf-core/rnaseq v3.18.0 - Lithium
+                   Lynx
+                  },
+  month        = dec,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {3.18.0},
+  doi          = {10.5281/zenodo.14537300},
+  url          = {https://doi.org/10.5281/zenodo.14537300},
+  swhid        = {swh:1:dir:ca630e51d1a8d26aa79ad417bc0dae393e6ef2af
+                   ;origin=https://doi.org/10.5281/zenodo.1400710;vis
+                   it=swh:1:snp:427f0da6e28ba02b98f0c8185177e78f21b5e
+                   049;anchor=swh:1:rel:072cd0fedea678e6ff3eacab2c708
+                   60d08b5e5b1;path=nf-core-rnaseq-b96a753
+                  },
+}
+@software{sarek,
+  author       = {Maxime U Garcia and
+                  Friederike Hanssen and
+                  Anders Sune Pedersen and
+                  Gisela Gabernet and
+                  WackerO and
+                  SusiJo and
+                  Adam Talbot and
+                  nf-core bot and
+                  Chela James and
+                  Bekir Ergüner and
+                  Alexander Peltzer and
+                  nickhsmith and
+                  José Fernández Navarro and
+                  Edmund Miller and
+                  Júlia Mir Pedrol and
+                  Francesco Lescai and
+                  Grant Neilson and
+                  Simon Pearce and
+                  Francisco Martínez and
+                  Pierre Lindenbaum and
+                  Nicolás Schcolnicov and
+                  David Mas-Ponte and
+                  Malin Larsson and
+                  Abhinav Sharma and
+                  Robert Syme and
+                  Chase Mateusiak and
+                  Famke Bäuerle and
+                  Lucia Conde and
+                  Lasse Folkersen and
+                  Adrian Lärkeryd},
+  title        = {nf-core/sarek: 3.5.0 - Áhkájiegna},
+  month        = dec,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {3.5.0},
+  doi          = {10.5281/zenodo.14412016},
[…]
