### 2025-01-10 Juke34/AliNe#4 @gchure
JOSS Review: Software Paper Comments
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/7545?notification_referrer_id=NT_kwDOAGdjK7MxMzY1NjU0MzgwMDo2Nzc1NTk1) and am opening issues as I progress through my checklist. This issue pertains to the software paper.

Overall, I think the paper is in good shape, but it could use some improvements and elaboration. Namely,

1. A diagram outlining the pipeline summary would be very helpful. You have this already in the README, but its inclusion and discussion in the software paper would be very helpful.

2. The paper lacks a discussion of the state-of-the-field with no citation/mention of other nextflow alignment tools, e.g. [nf-alignment](https://github.com/crukci-bioinformatics/nf-alignment), [alignment-nf](https://github.com/AndersenLab/alignment-nf), [sanger-tol/readmapping](https://github.com/sanger-tol/readmapping). I am not deeply enmeshed in this field, but was able to find other nf pipelines for alignment. The paper should more clearly state why you've written AliNe and what gaps in the current pipeline stack AliNe fills.

Other than that, I think the paper is in good shape. I will be opening other issues to discuss the software specifically.
