### 2025-01-10 Juke34/AliNe#5 @gchure
JOSS Review: Validity of Supported Methods
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/7545?notification_referrer_id=NT_kwDOAGdjK7MxMzY1NjU0MzgwMDo2Nzc1NTk1) and am opening issues as I progress through my checklist. This issue pertains an issue with supported alignment methods.

I admire the work you've put into supporting a broad range of alignment packages/algorithms, but I would appreciate some clarity on why you've chosen some and omitted others.

For example, you have not included support for [kallisto](https://pachterlab.github.io/kallisto/about) which uses pseudoalignment in quantification of RNAseq data. Is there a reason this is not included?

More importantly, I'm concerned by including Tophat as a valid alignment library at all. The corresponding author of Tophat has [repeatedly asked people to stop using the tool](https://x.com/lpachter/status/937055346987712512) as there are other, more advanced tools now.

In general, I think there should be some explanation of your process for deciding what is supported by AliNE.
