### 2025-01-10 #7545 @gchure
Hi @jromanowska, hope you're doing well! I wanted to give a quick update that I've opened some issues with #5 being the most substantive to the work. I'm having some installation issues wrt Docker on my machine that are stopping me from running the tests, but will work on that over the next few days.
