### 2025-01-10 Juke34/AliNe#4 @Juke34
Thank you for your valuable feedback.

1. I have included a diagram summerazing the pipeline steps
2.  In the last paragraph of the Statement of need I have added a discussion about the state-of-the-field. I Mention several pipeline including  [sanger-tol/readmapping](https://github.com/sanger-tol/readmapping). I have decided to not mention [nf-alignment](https://github.com/crukci-bioinformatics/nf-alignment) and [alignment-nf](https://github.com/AndersenLab/alignment-nf)  because I already mention 4 tools that are more known in the field. These tools have same limitation as the ones I mention, to know, specific to a type of bioinformatics analysis and for most of them specific to a sequencing technologies.

I had forgotten to re-compile the paper, you probably didn't seen these changes
