### 2025-01-10 Juke34/AliNe#5 @Juke34
Thank you for your review and valuable feedbacks.

I have tried to gather the tools I was aware of, and forgot kallisto. I didn't know it was possible to get sam/bam output. Reading the manual I realised my mistake. Thank you for pointing it out. I will try to add it in the next release. My longer-term goal is to gather all possible alignment softwares that provide sam/bam output.

I am aware of the concern with Topcoat but decided to add it with reproducibility in mind. Indeed I wanted to be able to rerun old analyses that used Tophat. With AliNe it will be easy to check and compare the contribution of new tools.
