### 2025-01-10 cf75192
fix citation list
