### 2025-01-10 #7545 @Juke34
This version take in consideration reviews from @rcannood and @gchure
