### 2025-01-11 Juke34/AliNe#8 @rcannood
JOSS Review: Review Documentation
I'm reviewing the JOSS submission of AliNe in https://github.com/openjournals/joss-reviews/issues/7545.

# Example usage

>  Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The Usage section in the readme currently contains:

* Check available options with `nextflow run ... --help`
* Command will look like that: `nextflow run ... -profile docker <rest of parameters>`.

Could you add an example of a command with actual input parameters?

Example:

```bash
nextflow run Juke34/AliNe \
  -r v1.0.0 \
  -profile docker \
  --reads https://github.com/Juke34/AliNe/raw/refs/heads/main/test/illumina/yeast_R1.fastq.gz \
  --genome https://raw.githubusercontent.com/Juke34/AliNe/refs/heads/main/test/yeast.fa \
  --read_type short_single \
  --aligner bbmap,bowtie2,bwaaln,bwamem,bwasw,graphmap2,hisat2,minimap2,ngmlr,nucmer,star,subread,sublong,tophat2 \
  --trimming_fastp \
  --star_options "--genomeSAindexNbases 9" \
  --multiqc_config https://raw.githubusercontent.com/Juke34/AliNe/refs/heads/main/config/multiqc_conf.yml
```

(The above is only possible after #7 is fixed)

or:

```bash
git clone https://github.com/Juke34/AliNe --depth=1

nextflow run Juke34/AliNe \
  -r v1.0.0 \
  -profile docker \
  --reads AliNe/test/illumina \
  --genome AliNe/test/yeast.fa \
  --read_type short_single \
  --aligner bbmap,bowtie2,bwaaln,bwamem,bwasw,graphmap2,hisat2,minimap2,ngmlr,nucmer,star,subread,sublong,tophat2 \
  --trimming_fastp \
  --star_options "--genomeSAindexNbases 9" \
  --multiqc_config AliNe/config/multiqc_conf.yml
```

This would greatly help users run the workflow :)
