### 2025-01-11 Juke34/AliNe#10 @rcannood
JOSS Review: Automated tests
I'm reviewing the JOSS submission of AliNe in https://github.com/openjournals/joss-reviews/issues/7545.

# Automated tests

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

Please add a GitHub action to the repository that runs the workflow on some test data after every commit and on pull requests.
