### 2025-01-13 Juke34/AliNe#5 @gchure
Got, thanks for the clarification. Glad you'll add kallisto -- it's a great tool.

As to Tophat, I understand why you may want to include it for rerunning old analyses, but I still don't think it should be supported at all. As this is open-source software being written to be used by many people with ease, I strongly believe that it shouldn't enable the use of outdated, inefficient methods that are no longer supported. Hisat2 is the successor to tophat, so I really don't see a reason to keep that as an enabled method.
