### 2025-01-13 Juke34/AliNe#11 @gchure
JOSS Review: Insufficient Container Information
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/7545?notification_referrer_id=NT_kwDOAGdjK7MxMzY1NjU0MzgwMDo2Nzc1NTk1) and am opening issues as I progress through my checklist. This issue pertains to a lack of clarity on containerization.

The installation instructions state that I will need either Docker or Singularity to use this software and points the user to the generic Docker and Singularity webpages. However, you include no dockerfile, singularity definition file, or prebuilt container for the user, which undercuts the point of containerization in the first place.

Please provide either a dockerfile/definition-file and/or upload a prebuilt container to Dockerhub/SylabsLibrary.
