### 2025-01-13 ae576ce
Fix #8 add a complete example how to run AliNe
