### 2025-01-14 dcfdbe3
add warning for Tophat2 to avoid usage outside of  comparisons and reproducibility of ancient analyses. #5
