### 2025-01-14 Juke34/AliNe#12 @gchure
JOSS Review: Contributor Guidelines
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/7545?notification_referrer_id=NT_kwDOAGdjK7MxMzY1NjU0MzgwMDo2Nzc1NTk1) and am opening issues as I progress through my checklist. This issue pertains to a lack of contribution guidelines.

In the `README` of your repository, please add guidelines for people to contribute to this project. You should address how people should i) open issues, ii) submit PRs, and iii) properly test modifications.
