### 2025-01-14 #7545 @gchure
Hi @jromanowska and @Juke34, I've finished what I can of the review thus far. Issue [#11](https://github.com/Juke34/AliNe/issues/11) is the most pressing as it is precluding me from testing the functionality of the software. More specific information (container/init file) is needed to run the software as advertised. Issue [#5](https://github.com/Juke34/AliNe/issues/5) is a scientific blocker. From my perspective, this software should not enable the use of an outdated method (i.e. TopHat) which has numerous, well documented issues.

Once #11 is addressed, I'll continue with my review.
