### 2025-01-15 d99e48e
add kallisto, rewrite how read length is handled (needed for kallisto) #5
