### 2025-01-15 Juke34/AliNe#5 @Juke34
I have added a warning when Tophat2 is used saying `Tophat2 has been deprecated. The developers recommend to switch to HISAT2.`

I have also added kallisto
