### 2025-01-15 8dc9ba2
point to the CONTRIBUTING.md from the README #12
