### 2025-01-15 Juke34/AliNe#11 @Juke34
I believe there might be a misunderstanding.
The only prerequisites are that you have Nextflow installed and  a container platform (docker or singularity/apptainer). There is no need for any additional definition file (such as Dockerfiles or Singularity definition files).
In the `Usage` section, there is a `Profile` paragraph where I mention `To run the workflow you must select a profile according to the container platform you want to use...` So if you want to use Docker, simply add `-profile docker` when calling AliNe, as shown in the examples provided in the README. If AliNe is ran without this information I throw an error telling the user: `No executer selected: -profile docker/singularity`.
Nextflow will handle the rest by retrieving the containers specified in the AliNe workflow from container registries. Each step of the pipeline will be executed in a dedicated container under the wood. If you’d like to review the containers being used, they are defined here: https://github.com/Juke34/AliNe/blob/main/config/softwares.config
