+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**General (please complete the following information):**
+- AliNe version (e.g. v1.0.1]
+- OS: [e.g. macOS, Ubuntu, CentOS]
+- Command ran
+
+**To Reproduce**
+Scripts and parameters to reproduce the behavior.
+Input file description
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots/sample of file to help explain your problem.
+
+**Additional context**
+Add any other context about the problem here.
+Reference any external link/discussion related to the issue (e.g. link to Biostars.org)
