+++ .github/workflows/main.yml
@@ -30,5 +30,6 @@ jobs:
         
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@v3
+      - name: test short paired
         run: nextflow run -profile docker,test_illumina_paired aline.nf
 
