+++ .github/workflows/main.yml
@@ -14,7 +14,7 @@ concurrency:
 jobs:  
 
   # This workflow contains a second job called "build2"
-  build_perl536:
+  test_aline:
     # avoid to run twice push and PR
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
 
