+++ .github/workflows/main.yml
@@ -30,7 +30,7 @@ jobs:
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
       - name: Install Nextflow dependencies
-        run: apk add --no-cache bash curl openjdk11
+        run: apt-get install openjdk-11-jdk
       - name: Install Nextflox
         run: curl -s https://get.nextflow.io | bash && mv nextflow /usr/local/bin && chmod +x /usr/local/bin/nextflow
         
