### 2025-01-16 2e6376e
fix cache key and path for java
