### 2025-01-16 a5626f2
add -ansi-log to get less output lines
