### 2025-01-16 Juke34/AliNe#5 @gchure
Thanks for adding Kallisto.

A warning does not sufficiently address my concern. This still makes it easy for people to use software that shouldn't be used. Tophat should not be included as a supported alignment method, full stop.
