### 2025-01-16 Juke34/AliNe#11 @gchure
Got it! Thanks. I was able to get the tests to run successfully locally.

A minor point -- `conda install nextflow` does not work without adding the bioconda channel. I would adjust the readme to run the following:

```
conda install bioconda::nextflow
```
