### 2025-01-17 Juke34/AliNe#11 @Juke34
Great!

good catch, I will add the bioconda channel.
