### 2025-01-17 402f7ac
add bioconda channel in conda installation of nextflow
