### 2025-01-17 535e735
add kallisto and bowtie1
