+++ paper/paper.bib
@@ -167,6 +167,23 @@
    url = {https://pubmed.ncbi.nlm.nih.gov/20080505/},
    year = {2010},
 }
+@article{bowtie,
+   abstract = {Bowtie is an ultrafast, memory-efficient alignment program for aligning short DNA sequence reads to large genomes. For the human genome, Burrows-Wheeler indexing allows Bowtie to align more than 25 million reads per CPU hour with a memory footprint of approximately 1.3 gigabytes. Bowtie extends previous Burrows-Wheeler techniques with a novel quality-aware backtracking algorithm that permits mismatches. Multiple processor cores can be used simultaneously to achieve even greater alignment speeds. Bowtie is open source http://bowtie.cbcb.umd.edu. © 2009 Langmead et al.; licensee BioMed Central Ltd.},
+   author = {Ben Langmead and Cole Trapnell and Mihai Pop and Steven L. Salzberg},
+   doi = {10.1186/GB-2009-10-3-R25/TABLES/5},
+   issn = {14747596},
+   issue = {3},
+   journal = {Genome Biology},
+   keywords = {Animal Genetics and Genomics,Bioinformatics,Evolutionary Biology,Human Genetics,Microbial Genetics and Genomics,Plant Genetics and Genomics},
+   month = {3},
+   pages = {1-10},
+   pmid = {19261174},
+   publisher = {BioMed Central},
+   title = {Ultrafast and memory-efficient alignment of short DNA sequences to the human genome},
+   volume = {10},
+   url = {https://genomebiology.biomedcentral.com/articles/10.1186/gb-2009-10-3-r25},
+   year = {2009},
+}
 @article{bowtie2,
    abstract = {As the rate of sequencing increases, greater throughput is demanded from read aligners. The full-text minute index is often used to make alignment very fast and memory-efficient, but the approach is ill-suited to finding longer, gapped alignments. Bowtie 2 combines the strengths of the full-text minute index with the flexibility and speed of hardware-accelerated dynamic programming algorithms to achieve a combination of high speed, sensitivity and accuracy. © 2012 Nature America, Inc. All rights reserved.},
    author = {Ben Langmead and Steven L. Salzberg},
@@ -183,6 +200,23 @@
    volume = {9},
    url = {https://pubmed.ncbi.nlm.nih.gov/22388286/},
    year = {2012},
+}
+@article{kallisto,
+   abstract = {A pseudoalignment-based method enables faster quantification and measurement of uncertainty in RNA-seq experiments. We present kallisto, an RNA-seq quantification program that is two orders of magnitude faster than previous approaches and achieves similar accuracy. Kallisto pseudoaligns reads to a reference, producing a list of transcripts that are compatible with each read while avoiding alignment of individual bases. We use kallisto to analyze 30 million unaligned paired-end RNA-seq reads in <10 min on a standard laptop computer. This removes a major computational bottleneck in RNA-seq analysis.},
+   author = {Nicolas L. Bray and Harold Pimentel and Páll Melsted and Lior Pachter},
+   doi = {10.1038/nbt.3519},
+   issn = {1546-1696},
+   issue = {5},
+   journal = {Nature Biotechnology 2016 34:5},
+   keywords = {Genome informatics,Software,Transcriptomics},
+   month = {4},
+   pages = {525-527},
+   pmid = {27043002},
+   publisher = {Nature Publishing Group},
+   title = {Near-optimal probabilistic RNA-seq quantification},
+   volume = {34},
+   url = {https://www.nature.com/articles/nbt.3519},
+   year = {2016},
 }
 @article{subread,
[…]
