### 2025-01-17 #7545 @Juke34
This version responds to the latest concerns of the reviewiers.
