### 2025-01-23 #7545 @rcannood
Thanks for the reminder -- it was quite busy the last week.

I'm currently reviewing the functionality claimed by authors and the documentation thereof, which are the last two items on my checklist.
