### 2025-01-23 Juke34/AliNe#18 @Juke34
The last release point to an ancient commit where tophat was still around. As soon as I fix the last concerns I will update the release appropriately.
I will answer the different problems you raised in separate issues:
