### 2025-01-24 31dec6d
#20 clean multiqc output to avoid duplicated name in the general statistics table, add section in readme to explain how to look at the results
