### 2025-01-25 3338c37
add output structure information
