+++ README.md
@@ -21,6 +21,7 @@ AliNe is a pipeline written in Nextflow that aims to efficiently align reads aga
         * [Singularity](#singularity)  
    * [Usage and test](#usage)
    * [Parameters](#parameters)
+   * [Output](#output)
    * [Contributing](#contributing)
 
 ## Foreword
@@ -349,6 +350,63 @@ On success you should get a message looking like this:
         --subread_options           additional options for subread
         --sublong_options           additional options for sublong
 ```
+
+## Output
+
+Here the description of typical ouput you will get from AliNe:  
+
+```
+└── alignment_results                                         # Output folder set using --outdir. Default: <alignment_results>
+    │
+    ├── fastp                                                 # Folder - trimming with fastp (optional - if trimming activated by the user)
+    │   ├── sample1_fastp_report.html                         # fastp report for sample1
+    │   └── sample1_seqkit_trim.fastq.gz                      # sample1 trimmed fastq file
+    │
+    ├── seqkit_score                                          # Folder containing Sequencing scoring system detected with Seqkit
+    │   └── sample1.result.txt                                # Information about scoring system detected in sample1 (Phred+33, Phred+64 and Solexa), and change applied
+    │
+    ├── mean_read_length                                      # Folder with mean read length computed in bash (optional - done if selected aligners need the info and no value provided by the user)
+    │   └── sample1_seqkit_trim_sampled_read_length.txt       # Mean read length for sample1
+    │
+    ├── salmon_libtype                                        # Librairy information (read orientation and strand information) detected via Salmon
+    │       └── sample1_lib_format_counts.json                # Librairy information detectected for sample1
+    |
+    ├── alignment                                             # Folder gathering all alignment output (indicies, sorted bam and logs)
+    │   ├── aligner1                                          # Folder gathering data produced by aligner 
+    │   │   ├── indicies                                      # Contains the genome index for the aligner
+    │   │   │   └── ...                                       #
+    │   │   ├── sample1_seqkit_trim_aligner1_sorted.log       # Ccontains the log of the aligner
+    │   │   └── sample1_seqkit_trim_aligner1_sorted.bam       # Sorted bam output
+    │   └── aligner2                                          # Folder gathering data produced by aligner 
+    │       ├── indicies                                      # Contains the genome index for the aligner
+    │       │   └── ...                                       # 
+    │       ├── sample1_seqkit_trim_aligner2_sorted.log       # Contains the log of the aligner
+    │       └── sample1_seqkit_trim_aligner2_sorted.bam       # Sorted bam output
+    │
+    ├── fastqc                                                # FastQC statistics folder
+    │   ├── raw                                               # Folder with FastQC result for raw data
+    │   │   └── fastqc_sample1_raw_logs                       # Folder with FastQC result for raw sample1 data
+    │   │       ├── sample1_fastqc.html                       # FastQC interactive file summarizing the results of the analysis, with graphs and interpretations.
+    │   │       └── sample1_fastqc.zip                        # Contains all the detailed data and graphics generated by FastQC
+    │   └── trimming_fastp                                    # Folder with FastQC result for trimmed data (optional - if trimming activated by the user)
+    │   │   └── fastqc_sample1_trimmed_logs                   # FastQC output folder for trimmed sample1 data
+    │   │       ├── sample1_seqkit_trim_fastqc.html           # FastQC interactive file summarizing the results of the analysis, with graphs and interpretations.
+    │   │       └── sample1_seqkit_trim_fastqc.zip            # Contains all the detailed data and graphics generated by FastQC
+    │   ├── aligner1                                                 # FastQC output folder for data aligned with aligner1 
[…]
