+++ .github/workflows/main.yml
@@ -19,7 +19,7 @@ jobs:
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
 
     # The type of runner that the job will run on
-    runs-on: ubuntu:24.04
+    runs-on: ubuntu-24.04
     services:
       docker:
         image: docker:dind
