### 2025-01-28 e78decc
#20 add citationt to point to documentation about daisy-chaining
