+++ paper/paper.bib
@@ -1,3 +1,9 @@
+@Online{nfcascade,
+  accessed = {2025-01-24},
+  author   = {{Mahesh Binzer-Panchal}},
+  title    = {A proof of concept daisy-chaining Nextflow workflows},
+  url      = {https://github.com/mahesh-panchal/nf-cascade/},
+}
 @software{tronflow,
   author       = {Pablo Riesgo-Ferreiro and
                   tronbfxdev and
