+++ paper/paper.md
@@ -27,7 +27,7 @@ AliNe supports short reads (both paired-end and single-end) as well as long read
 The complexity of sequencing data analysis, especially with the advent of next-generation sequencing (NGS), has created a demand for robust, scalable, and reproducible workflows that can handle large datasets while allowing users to flexibly choose their preferred analysis tools. Numerous bioinformatics analyses rely on read alignment steps that can play a major role in the quality of the downstream results. Each alignment tool may behave differently and have its own strengths. Users may need to try out several alignment tools in order to achieve optimal results specific to a project's requirements. AliNe empowers researchers with a single-entry point for alignment, eliminating the need to juggle between disparate tools. By integrating multiple aligners into a unified pipeline, AliNe simplifies the process of alignment algorithm selection and establishes a standardized workflow that ensures both reproducibility and transparency across analyses
 
 The majority of pipelines integrating read alignment steps are specialized for particular sequencing technologies and types of analyses [@tronflow] [@nfcorernaseq] [@sarek] [@readmapping]. There remains a lack of unified, versatile workflows capable of integrating aligners into a single, user-friendly platform.
-AliNe addresses this challenge by providing a comprehensive, streamlined solution for alignment tasks across a wide range of sequencing technologies, whatever the field of application. Moreover, AliNe can be integrated as a sub-workflow into other Nextflow pipeline and provide a standardize step that offer users a wider choice of alignment tools.
+AliNe addresses this challenge by providing a comprehensive, streamlined solution for alignment tasks across a wide range of sequencing technologies, whatever the field of application. Moreover, AliNe can be integrated as a sub-workflow into other Nextflow pipeline [@nfcascade] and provide a standardize step that offer users a wider choice of alignment tools.
 
 # Workflow Overview
 
