### 2025-01-28 f37abbd
#20 add samtools stats and provide more information about statistics
