### 2025-02-05 08efdde
add ref daisy chaining
