+++ paper/paper.bib
@@ -1,6 +1,14 @@
 @Online{nfcascade,
   accessed = {2025-01-24},
+  year     = 2024,
   author   = {{Mahesh Binzer-Panchal}},
+  title    = {A proof of concept daisy-chaining Nextflow workflows},
+  url      = {https://github.com/mahesh-panchal/nf-cascade/},
+}
+@Online{daisychainingaline,
+  accessed = {2025-02-03},
+   year     = 2025,
+  author   = {{Jacques Dainat}},
   title    = {A proof of concept daisy-chaining Nextflow workflows},
   url      = {https://github.com/mahesh-panchal/nf-cascade/},
 }
