### 2025-03-01 #7545 @Juke34
Hi @jromanowska,

The JOSS release of AliNe is [v1.1.1](https://github.com/Juke34/AliNe/tree/v1.1.1) and the corresponding ZENODO release is [DOI: 10.5281/zenodo.14953389](https://doi.org/10.5281/zenodo.14953389).
