### 2025-03-03 8e71057
Update paper.md
some grammar corrections and suggestions for re-phrasing
