### 2025-03-03 #7545 @Juke34
Hi @jromanowska I edited the zenodo side to be similar.
