### 2025-03-03 #7545 @Juke34
Should I update the release and the zenodo to integrate your last changes?
