### 2025-03-03 #7545 @jromanowska
> Should I update the release and the zenodo to integrate your last changes?

No, don't update the release and zenodo - these were only changes in the paper text.
