### 2025-03-04 #7545 @jromanowska
@Juke34 - I've noticed couple of small mistakes in references:
1. "Twelve years of SAMtools and BCFtools" lacks the final author (if you compare with [the webpage](https://academic.oup.com/gigascience/article/10/2/giab008/6137722))
2. [`nf-cascade`](https://github.com/mahesh-panchal/nf-cascade/) is listed as authored by you, but I can see it's done by another author
_there are two references to the same github repo (`@nfcascade` and `@daisychainingaline`) - could you check that and fix?_
4. "Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM" is a manuscript title while it's formatted as journal title?
5. correct the capitalization in: "Fast and accurate short read alignment with Burrows–Wheeler transform", "The subreadr aligner:", "SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation"
6. "Salmon provides fast and bias-aware quantification of transcript expression" and "Accurate detection of complex structural variations using single-molecule sequencing" - the journal title should say only "Nature Methods"

Please, correct and re-generate the pdf here.
