+++ paper/paper.bib
@@ -9,8 +9,8 @@
   accessed = {2025-02-03},
    year     = 2025,
   author   = {{Jacques Dainat}},
-  title    = {A proof of concept daisy-chaining Nextflow workflows},
-  url      = {https://github.com/mahesh-panchal/nf-cascade/},
+  title    = {Daisy chaining AliNe},
+  url      = {https://github.com/Juke34/daisy-chaining-AliNe},
 }
 @software{tronflow,
   author       = {Pablo Riesgo-Ferreiro and
@@ -287,7 +287,7 @@
    month = {10},
    pmid = {27706213},
    publisher = {Public Library of Science},
-   title = {SeqKit: A cross-platform and ultrafast toolkit for FASTA/Q file manipulation},
+   title = {SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation},
    volume = {11},
    year = {2016},
 }
@@ -310,7 +310,7 @@
    doi = {10.1038/nmeth.4197},
    issn = {1548-7105},
    issue = {4},
-   journal = {Nature Methods 2017 14:4},
+   journal = {Nature Methods},
    keywords = {Software,Statistical methods,Transcriptomics},
    month = {3},
    pages = {417-419},
@@ -354,7 +354,7 @@
    doi = {10.1038/s41592-018-0001-7},
    issn = {1548-7105},
    issue = {6},
-   journal = {Nature Methods 2018 15:6},
+   journal = {Nature Methods},
    keywords = {Genome informatics,Software,Structural variation},
    month = {4},
    pages = {461-468},
@@ -413,7 +413,7 @@
 }
 @article{samtools,
    abstract = {Background: SAMtools and BCFtools are widely used programs for processing and analysing high-throughput sequencing data. They include tools for file format conversion and manipulation, sorting, querying, statistics, variant calling, and effect analysis amongst other methods. Findings: The first version appeared online 12 years ago and has been maintained and further developed ever since, with many new features and improvements added over the years. The SAMtools and BCFtools packages represent a unique collection of tools that have been used in numerous other software projects and countless genomic pipelines. Conclusion: Both SAMtools and BCFtools are freely available on GitHub under the permissive MIT licence, free for both non-commercial and commercial use. Both packages have been installed >1 million times via Bioconda. The source code and documentation are available from https://www.htslib.org.},
-   author = {Petr Danecek and James K. Bonfield and Jennifer Liddle and John Marshall and Valeriu Ohan and Martin O. Pollard and Andrew Whitwham and Thomas Keane and Shane A. McCarthy and Robert M. Davies},
+   author = {Petr Danecek, James K Bonfield, Jennifer Liddle, John Marshall, Valeriu Ohan, Martin O Pollard, Andrew Whitwham, Thomas Keane, Shane A McCarthy, Robert M Davies, Heng Li},
    doi = {10.1093/gigascience/giab008},
    issn = {2047217X},
    issue = {2},
