### 2025-03-04 #7545 @Juke34
Thank you for your carefull review

> [@Juke34](https://github.com/Juke34) - I've noticed couple of small mistakes in references:
>
> 1. "Twelve years of SAMtools and BCFtools" lacks the final author (if you compare with [the webpage](https://academic.oup.com/gigascience/article/10/2/giab008/6137722))

What a shame to have forgotten Heng Li, thank you for pointing it is fixed now.

> 2. [`nf-cascade`](https://github.com/mahesh-panchal/nf-cascade/) is listed as authored by you, but I can see it's done by another author
>    _there are two references to the same github repo (`@nfcascade` and `@daisychainingaline`) - could you check that and fix?_

I fixed the mistake.

> 3. "Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM" is a manuscript title while it's formatted as journal title?

I'm not sure to understand why you say it is formatted as journal title... is it due to BWA-MEM in capital? It is just kept as provided [here]("Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM" is a manuscript title while it's formatted as journal title?) or [here](https://github.com/lh3/bwa).
Let me know what you expect (lower case "BWA-MEM"?).

> 4. correct the capitalization in: "Fast and accurate short read alignment with Burrows–Wheeler transform", "The subreadr aligner:", "SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation"

I do not see any rule from JOSS saying we need to capitalize title, I thought we must stick to the original capitalisation where it was published. In that light, I updated only the last one (SeqKit). Let me know if I really have to capitalize the two others.

> 5. "Salmon provides fast and bias-aware quantification of transcript expression" and "Accurate detection of complex structural variations using single-molecule sequencing" - the journal title should say only "Nature Methods">

Good catch, it is fixed now.

> Please, correct and re-generate the pdf here.
