+++ paper/paper.bib
@@ -248,7 +248,7 @@
    url = {https://pubmed.ncbi.nlm.nih.gov/23558742/},
    year = {2013},
 }
-@misc{bwamem,
+@article{bwamem,
       title={Aligning sequence reads, clone sequences and assembly contigs with BWA-MEM}, 
       author={Heng Li},
       abstract = {Summary: BWA-MEM is a new alignment algorithm for aligning sequence reads or long query sequences against a large reference genome such as human. It automatically chooses between local and end-to-end alignments, supports paired-end reads and performs chimeric alignment. The algorithm is robust to sequencing errors and applicable to a wide range of sequence lengths from 70bp to a few megabases. For mapping 100bp sequences, BWA-MEM shows better performance than several state-of-art read aligners to date. Availability and implementation: BWA-MEM is implemented as a component of BWA, which is available at http://github.com/lh3/bwa. Contact: hengli@broadinstitute.org},
