### 2025-03-04 #7545 @Juke34
> Thanks for quick fixes!
>
> What I meant in pt.4 above is that the "Burrows-Wheeler" and "FASTA/Q" are written as "burrows-wheeler" and "FASTA/q" in your reference list.

For that point I don't know what I should do. The JOSS reference processing put everything in lower case except first character in the title and any word that is in CAPITAL it is whole. So DNA stay DNA but "Burrows-Wheeler" becomes "burrows-wheeler". `/` is seen as non-word character and "FASTA/Q" becomes "FASTA/q".

> Ad.point 3 - I ment that there is no journal title and the article title is in italics, as if it was journal title. I can see in the .bib file that you've formatted this citation from arxiv as "misc" while another one, from biorxiv is correctly formatted as "article". If you change the format to "article", it will look nicer.

This is fixed now
