### 2025-03-04 #7545 @jromanowska
Oh, it's the bibtex-way: you put the entire title in extra pair of curly braces :) I've made a pull request for you.
