### 2025-03-05 #7545 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
