+++ paper/paper.md
@@ -32,7 +32,7 @@ Furthermore, leveraging the concept of daisy-chaining Nextflow workflows [@nfcas
 # Workflow Overview
 
 AliNe takes sequencing reads in FASTQ format and aligns them to a user-supplied reference genome in FASTA format, using user-selected aligners.
-The workflow is fully configurable to accommodate various data types and gives users the ability to fine-tune tool alignment parameters independently. Nevetheless, AliNe is designed to minimize user inputs and avoid common mistakes like the definition of scoring system used by the sequencing machine (Phred+33, Phred+64, and Solexa) or, in the case of RNAseq data, the strandedness (stranded or unstranded) or the orientation in case of paired-end reads (forward reverse i.e., fr; reverse forward i.e., rf; forward forward i.e., ff). If not provided by the user, AliNe will deduce this information and adapt the option of the different downstream aligners accordingly. AliNe suggests also a proper aligner, depending on the input data provided.
+The workflow is fully configurable to accommodate various data types and gives users the ability to fine-tune tool alignment parameters independently. Nevertheless, AliNe is designed to minimize user inputs and avoid common mistakes like the definition of scoring system used by the sequencing machine (Phred+33, Phred+64, and Solexa) or, in the case of RNAseq data, the strandedness (stranded or unstranded) or the orientation in case of paired-end reads (forward reverse i.e., fr; reverse forward i.e., rf; forward forward i.e., ff). If not provided by the user, AliNe will deduce this information and adapt the option of the different downstream aligners accordingly. AliNe suggests also a proper aligner, depending on the input data provided.
 
 To run AliNe you need to select one or several aligners and specify the type of input data: short single-end reads, short paired-end reads, long reads generated by PacBio or long reads generated by Oxford Nanopore Technologies (ONT).
 AliNe integrates an optional read trimming step using Fastp [@fastp]. 
@@ -45,7 +45,7 @@ The pipeline supports parallel execution, ensuring efficient processing of large
 
 The pipeline is modular, enabling future expansion and the integration of additional tools or customization of downstream analysis steps. Users can adapt AliNe to their specific research needs, whether for RNA-seq, ChIP-seq, or whole-genome resequencing projects.
 
-![Overview of the different steps inplemented in AliNe.](aline_overview.png)
+![Overview of the different steps implemented in AliNe.](aline_overview.png)
 
 # Availability
 
